fix: better file organization for backend crate
