+++ crates/backend/src/services/mod.rs
@@ -1,2 +1,4 @@
-pub mod compile;
+mod compile;
 mod sandbox;
+
+pub use compile::{route_compile, CompilationRequest, CompilationResult};
